pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.13.4",
    "langdetect>=1.0.9",
    "openai>=1.91.0",
    "pandas>=2.3.0",
    "pymupdf>=1.26.1",
    "pypdf2>=3.0.1",
    "python-dateutil>=2.9.0.post0",
    "pyyaml>=6.0.2",
    "requests>=2.32.4",
    "tqdm>=4.67.1",
]
